Simplify sending of initial close packets
